network: deleteContractId API 연결 - 2
